Add TLS renegotiation test
Michael Boman ported sslscan to Windows. He also added a basic reneg test.
This integrates his test and some Tor OpenSSL magic to attempt TLS
renegotiation.  It does not seem to work with encrypted.google.com:443 but it
does with riseup.net:443 and this is annoying. It appears to be the case that
OpenSSL 0.9.8k 25 Mar 2009 cannot deal with the TLS secure renegotiation
extension released in late 2009. A later version of OpenSSL may work with all
of the popular SSL/TLS servers and more testing is required.