Proper fix for the missing EXPORT and 56bit SSLv2 ciphers. Fixes #77
The previous fix of using enable-ssl2 and enable-weak-ciphers was
incomplete, because the 40 and 56bit SSLv2 ciphers were still disabled.
There's no option to re-enable them - they're commented out with #if 0
statements.

Building against a previous tag will mean missing out on future
features, so as a workaround the Makefile now replaces those #if 0
statements with #if 1 to re-enable the ciphers.

Bit of a bodge, but doesn't seem to break anything, and will allow us to
keep building against the latest version.