Fix TLS versions in heartbleed test
The fix for issue #38 isn't quite right since it only updated the
version in the TLS record layer and not the version in the client hello.

The client hello version is the one that actually influences the TLS
version negotiation. For the initial client hello, the tls record layer
version should be set to TLS 1.0