Unify configuration for OpenSSL.
MacOS isn't special and can properly build with the same arguments as the
"normal" build path.
While we are at it, go ahead and use Configure, which is the documented way to
invoke the OpenSSL build system (./config just calls Configure).

Tested on:	MacOS (arm64), FreeBSD (amd64, arm64), Ubuntu (aarch64)